Add apache math lib and trajectory file (#5)
I added the apache math library for polynomial fitting purposes and the trajectory file for aiming purposes. I edited build.gradle to include the JAR library, and it should be usable anywhere. Feel free to add more external libraries to the 'libs' folder if necessary, they should be able to be imported correctly.